WIP: Replace `chunk` module
This replaces the `chunk` module (with `BLTEFile`) with a chunk encoding
parser, which can operate on `Read` and `AsyncReadExt` traits rather
than `Vec<u8>`. This reduces memory usage.

The full BLTE file parser will be moved into a new file, with parsers
that use those traits.

* `CompressionMode` replaced with `ChunkEncodingHeader`. This now also
  parses the encrypted chunk header.

* Moved chunk encoding header parsing from `ChunkData` to
  `ChunkEncodingHeader`.

This still won't build as-is.